requirements.txt
flask
requests
python-dotenv
pyyaml
pytest
gunicorn
